requirements.txt
Flask
google-generativeai>=0.8.3
gunicorn
python-dotenv
requests